// src/types.ts
export type RootStackParamList = {
  Welcome: undefined;
  TermsOfServiceModal: undefined;
  // Signin: undefined;
  CreateAccount: undefined;
  Signin: { expoPushToken: string };
  // CreateAccount: { expoPushToken?: string | undefined };
  AccountCreated: undefined;
  Confirmation: undefined;
  ProfileSetup: { userId: string };
  EventlistTab: { userId?: string };
  Top: { userId?: string };
  AdminPage: { userId?: string };
  ReportPage:{ userId: string };
  QRCodeComponent: undefined;
  SearchUser: undefined;
  ImageUploadPage: { userId: string };
  Social: { userId: string };
  ChatList: undefined;
  ChatRoom: { chatRoomId: string; chatRoomTitle: string; activeTab: 'group' | 'personal';hostUserId?: string;receiverUserId?: string; };
  Home: undefined;
  alert: undefined;
  UserPage: undefined;
  ProfileEdit: { userId: string };
  AttributeSettings: { userId: string };
  TalkScreen: undefined;
  TermsOfServiceScreen: undefined; 
  SetHobbies: { isEdit: boolean };
  CreateArticleScreen:undefined;
  ArticleDetail: { articleId: string };
  ScheduleDetail:{ userId: string };
  NotificationListScreen:{ userId: string };
};